requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
requests
anyio
starlette
h11